Добавляет ветку с модулями
